+++ cedar-policy-validator/src/typecheck/test_optional_attributes.rs
@@ -20,17 +20,19 @@
 // GRCOV_STOP_COVERAGE
 
 use cedar_policy_core::{
-    ast::{BinaryOp, EntityUID, Expr, StaticPolicy, Var},
+    ast::{Expr, StaticPolicy, Var},
     parser::parse_policy,
 };
 use smol_str::SmolStr;
 
 use crate::{
     type_error::TypeError, types::EntityLUB, AttributeAccess, NamespaceDefinition,
-    NamespaceDefinitionWithActionAttributes,
+    NamespaceDefinitionWithActionAttributes, ValidationWarningKind,
 };
 
-use super::test_utils::{assert_policy_typecheck_fails, assert_policy_typechecks};
+use super::test_utils::{
+    assert_policy_typecheck_fails, assert_policy_typecheck_warns, assert_policy_typechecks,
+};
 
 fn schema_with_optionals() -> NamespaceDefinition {
     serde_json::from_str::<NamespaceDefinition>(
@@ -776,30 +778,17 @@ fn action_attrs_failing() {
         )],
     );
 
-    // Doesn't fail do to imprecision in ActionEntity LUB computation requiring `may_have_attr` to return true for ActionEntity types
+    // Doesn't fail due to imprecision in ActionEntity LUB computation requiring `may_have_attr` to return true for ActionEntity types
 
     let failing_policy = parse_policy(
         Some("0".to_string()),
         r#"permit(principal, action == Action::"view", resource) when { action has "" };"#,
     )
     .expect("Policy should parse.");
-    assert_policy_typecheck_fails(
+    assert_policy_typecheck_warns(
         schema.clone(),
         failing_policy,
-        vec![TypeError::impossible_policy(Expr::and(
-            Expr::and(
-                Expr::and(
-                    Expr::val(true),
-                    Expr::binary_app(
-                        BinaryOp::Eq,
-                        Expr::var(Var::Action),
-                        Expr::val(EntityUID::with_eid_and_type("Action", "view").unwrap()),
-                    ),
-                ),
-                Expr::val(true),
-            ),
-            Expr::has_attr(Expr::var(Var::Action), "".into()),
-        ))],
+        vec![ValidationWarningKind::ImpossiblePolicy],
     );
 
     let failing_policy = parse_policy(